requirements.txt
PyQt5==5.15.9
kivymd==1.1.1
pywin32==306
yt-dlp>=2023.12.30
ffmpeg-python>=0.2.0
python-ffmpeg>=2.0.0
beautifulsoup4==4.12.2
requests==2.31.0
moviepy==1.0.3
Pillow>=9.5.0
python-dotenv>=0.19.0
SQLAlchemy==2.0.25
pystray>=0.19.4
keyboard>=0.13.5
pyperclip>=1.8.2
readability-lxml>=0.8.1
html2text>=2020.1.16
beautifulsoup4>=4.9.3
tweepy>=4.12.0
instaloader>=4.9.5
python-instagram-api>=0.3.2

# Sosyal Medya Entegrasyonu
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-twitter>=3.5
instabot>=0.117.0
tiktok-api>=0.1.1
spotipy>=2.19.0
python-wordpress-xmlrpc>=2.3
blogger3>=0.0.3
medium-api>=0.5.0
linkedin-api>=2.0.0
facebook-sdk>=3.1.0
requests-oauthlib>=1.3.0
